Sort elements, rename workspace
Rename workspace to workspaceViewer, add settings option for removing hidden items, sort elements in workspace viewer, fixed some typings
