+++ package.json
@@ -718,6 +718,11 @@
           "type": "integer",
           "default": 8,
           "description": "Delay in milliseconds before sending each line to rterm (only applies if r.bracketedPaste is false)"
+        },
+        "r.workspaceViewer.removeHiddenItems" : {
+          "type" : "boolean",
+          "default" : false,
+          "description": "Remove hidden items when clearing workspace."
         }
       }
     }
